# river-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the River self-hosted media platform.\nAuth flow: POST /api/auth/login → use the returned\naccess_token as a Bearer token on the protected routes.\nThe stream_token is a longer-lived JWT restricted to\n/stream and /download endpoints (the \u003cvideo\u003e element\ncan't send custom auth headers; it goes in the URL as\n?token=…).",
        "title": "River API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "List service log entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by level (info|warn|error)",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by service name",
                        "name": "service",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO8601 lower bound",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO8601 upper bound",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{logs, total}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/logs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Append service log entry",
                "parameters": [
                    {
                        "description": "Log entry",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createLogReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/movies/{id}/subtitles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtitles"
                ],
                "summary": "List movie subtitles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Subtitle"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/subtitles": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtitles"
                ],
                "summary": "Create subtitle",
                "parameters": [
                    {
                        "description": "Subtitle metadata",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createSubtitleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Subtitle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/subtitles/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "subtitles"
                ],
                "summary": "Delete subtitle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subtitle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/subtitles/{id}/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "text/vtt"
                ],
                "tags": [
                    "subtitles"
                ],
                "summary": "Stream subtitle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subtitle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stream JWT",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tvshows/{id}/seasons/{seasonId}/episodes/{episodeId}/subtitles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtitles"
                ],
                "summary": "List episode subtitles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TV show ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Season ID",
                        "name": "seasonId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Episode ID",
                        "name": "episodeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Subtitle"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createLogReq": {
            "type": "object",
            "required": [
                "level",
                "message",
                "service"
            ],
            "properties": {
                "level": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "handlers.createSubtitleRequest": {
            "type": "object",
            "required": [
                "file_path",
                "language",
                "media_id",
                "media_type"
            ],
            "properties": {
                "file_path": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "media_id": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                }
            }
        },
...
        "models.Subtitle": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "description": "human-readable e.g. \"English\"",
                    "type": "string"
                },
                "language": {
                    "description": "BCP-47 tag e.g. \"en\"",
                    "type": "string"
                },
                "media_id": {
                    "type": "string"
                },
                "media_type": {
                    "description": "\"movie\" or \"episode\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT — paste \"Bearer \u003ctoken\u003e\" or use the Authorize button.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}